+++ runtime/frequency/src/tests/mod.rs
@@ -1,2 +1,3 @@
+pub mod mock;
+
 mod fee_estimation;
-mod mock;
